+++ src/cli/__init__.py
@@ -51,6 +51,7 @@ from ._common import (
 )
 from ._doctor_checks import (
     CheckLevels,
+    check_1_9_sed,
     check_docker_root_free_space,
     check_memory,
     check_mount_tmp,
@@ -367,9 +368,12 @@ def create(
             build_artifact=artifact_path,
             workspace_dir=workspace_dir,
         )
-    except subprocess.CalledProcessError:
+    except subprocess.CalledProcessError as exc:
         rmtree(path=str(workspace_dir), ignore_errors=True)
         click.echo(doctor_message)
+        click.echo()
+        click.echo('Original error:')
+        click.echo(exc.stderr)
         raise
 
     if enterprise:
@@ -940,6 +944,7 @@ def doctor() -> None:
     Diagnose common issues which stop DC/OS E2E from working correctly.
     """
     check_functions = [
+        check_1_9_sed,
         check_docker_root_free_space,
         check_memory,
         check_mount_tmp,