+++ src/cli/_doctor_checks.py
@@ -4,6 +4,7 @@ Checks for showing up common sources of errors with the Docker backend.
 
 import shutil
 import subprocess
+import tempfile
 from enum import IntEnum
 from pathlib import Path
 from tempfile import gettempdir, gettempprefix
@@ -312,7 +313,32 @@ def link_to_troubleshooting() -> CheckLevels:
         'If you continue to experience problems, more information is '
         'available at '
         'http://dcos-e2e.readthedocs.io/en/latest/docker-backend.html#troubleshooting'  # noqa: E501
+        '.'
     )
 
     _info(message=message)
+    return CheckLevels.NONE
+
+
+def check_1_9_sed() -> CheckLevels:
+    """
+    Warn if the system's version of ``sed`` is incompatible with legacy DC/OS
+    installers.
+    """
+    temp = tempfile.NamedTemporaryFile()
+    Path(temp.name).write_text('a\na')
+    sed_args = "sed '0,/a/ s/a/b/' " + temp.name
+    result = subprocess.check_output(args=sed_args, shell=True)
+
+    if result != b'b\na':
+        message = (
+            'The version of ``sed`` is not compatible with installers for '
+            'DC/OS 1.9 and below. '
+            'See '
+            'http://dcos-e2e.readthedocs.io/en/latest/versioning-and-api-stability.html#dc-os-1-9-and-below'  # noqa: E501
+            '.'
+        )
+        _warn(message=message)
+        return CheckLevels.WARNING
+
     return CheckLevels.NONE