+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Changelog
 Next
 ----
 
+- Add a new ``dcos-docker doctor`` check for suitable ``sed`` for DC/OS 1.9.
+
 2018.04.02.0
 ------------